+++ src/arch/riscv/riscv_common.c
@@ -60,7 +60,7 @@ void riscv_exception_handler(uint32_t mcause, uint32_t mepc, uint32_t mtval,
      * could fault (current, backtrace).  These three values come from
      * CSRs passed in registers and are always safe. */
     klogf("TRAP: cause=%lx mepc=%lx mtval=%lx sp=%lx\n",
-          mcause, mepc, mtval, saved_sp);
+          (unsigned long)mcause, (unsigned long)mepc, (unsigned long)mtval, (unsigned long)saved_sp);
 
     CRASH_LOG[0] = 0xDEADBEEFu;
     CRASH_LOG[1] = mcause;
@@ -68,8 +68,8 @@ void riscv_exception_handler(uint32_t mcause, uint32_t mepc, uint32_t mtval,
     CRASH_LOG[3] = mtval;
 
     klogf("  pid=%lu fp=%lx\n",
-          current ? (uint32_t)current->pid : 0xFFFFFFFFu,
-          saved_fp);
+          (unsigned long)(current ? (uint32_t)current->pid : 0xFFFFFFFFu),
+          (unsigned long)saved_fp);
 
     /* Walk the frame pointer chain from the faulting context.
      * Each frame: [fp-4] = ra, [fp-8] = prev fp.
@@ -79,7 +79,7 @@ void riscv_exception_handler(uint32_t mcause, uint32_t mepc, uint32_t mtval,
     for (uint32_t depth = 0; depth < 16 && fp >= 0x80000000u; depth++) {
         uintptr_t ra = *(uintptr_t *)(fp - 4);
         uintptr_t prev_fp = *(uintptr_t *)(fp - 8);
-        klogf("    #%u ra=%lx fp=%lx\n", (uint32_t)depth, (uint32_t)ra, (uint32_t)fp);
+        klogf("    #%u ra=%lx fp=%lx\n", (unsigned)depth, (unsigned long)ra, (unsigned long)fp);
         if (prev_fp <= fp) break;
         fp = prev_fp;
     }
