+++ src/kernel/mm/page.c
@@ -122,24 +122,29 @@ void mm_init(void) {
 
   /* ── Boot-time memory map ─────────────────────────────────────────────── */
   klog("MM: memory map\n");
-  klogf("MM:   kernel  %lx-%lx  %lu KB reserved\n", SRAM_KERNEL_BASE,
-        SRAM_KERNEL_BASE + SRAM_KERNEL_SIZE - 1u, SRAM_KERNEL_SIZE / 1024u);
+  klogf("MM:   kernel  %lx-%lx  %lu KB reserved\n",
+        (unsigned long)SRAM_KERNEL_BASE,
+        (unsigned long)(SRAM_KERNEL_BASE + SRAM_KERNEL_SIZE - 1u),
+        (unsigned long)(SRAM_KERNEL_SIZE / 1024u));
   if (stack_top > bss_end)
-    klogf("MM:     .data/.bss:  %lx B used, %lx B to stack top\n", kern_used,
-          stack_top - bss_end);
+    klogf("MM:     .data/.bss:  %lx B used, %lx B to stack top\n",
+          (unsigned long)kern_used, (unsigned long)(stack_top - bss_end));
   else
-    klogf("MM:     .data/.bss:  %lx B used\n", kern_used);
+    klogf("MM:     .data/.bss:  %lx B used\n", (unsigned long)kern_used);
 
   uintptr_t actual_base =
       (free_top > 0) ? (uintptr_t)free_stack[0] : pool_base;
-  klogf("MM:   pages   %lx-%lx %lu KB (%u x 4 KB, all free)\n", actual_base,
-        pool_base + page_count * PAGE_SIZE - 1u,
-        free_top * PAGE_SIZE / 1024u, free_top);
+  klogf("MM:   pages   %lx-%lx %lu KB (%u x 4 KB, all free)\n",
+        (unsigned long)actual_base,
+        (unsigned long)(pool_base + page_count * PAGE_SIZE - 1u),
+        (unsigned long)(free_top * PAGE_SIZE / 1024u), free_top);
 #if !defined(__m68k__) && !defined(__xtensa__) && !defined(__ia16__)
-  klogf("MM:   io_buf  %lx-%lx  %lu KB\n", SRAM_IOBUF_BASE,
-        SRAM_IOBUF_BASE + SRAM_IOBUF_SIZE - 1u, SRAM_IOBUF_SIZE / 1024u);
-  klogf("MM:   dma     %lx-%lx  %lu KB\n", SRAM_DMA_BASE,
-        SRAM_DMA_BASE + SRAM_DMA_SIZE - 1u, SRAM_DMA_SIZE / 1024u);
+  klogf("MM:   io_buf  %lx-%lx  %lu KB\n", (unsigned long)SRAM_IOBUF_BASE,
+        (unsigned long)(SRAM_IOBUF_BASE + SRAM_IOBUF_SIZE - 1u),
+        (unsigned long)(SRAM_IOBUF_SIZE / 1024u));
+  klogf("MM:   dma     %lx-%lx  %lu KB\n", (unsigned long)SRAM_DMA_BASE,
+        (unsigned long)(SRAM_DMA_BASE + SRAM_DMA_SIZE - 1u),
+        (unsigned long)(SRAM_DMA_SIZE / 1024u));
 #endif
 
 #ifdef PPAP_TESTS
@@ -238,7 +243,7 @@ static void stack_backtrace(void) {
   for (uint32_t depth = 0; depth < 16 && fp; depth++) {
     uintptr_t ra = *(uintptr_t *)(fp - 4);
     uintptr_t prev_fp = *(uintptr_t *)(fp - 8);
-    klogf("    #%u ra=%lx fp=%lx\n", depth, (uint32_t)ra, (uint32_t)fp);
+    klogf("    #%u ra=%lx fp=%lx\n", depth, (unsigned long)ra, (unsigned long)fp);
     if (prev_fp <= fp) break; /* stack grows down — prev fp must be higher */
     fp = prev_fp;
   }
@@ -262,8 +267,8 @@ void page_free(void *page) {
   for (uint32_t i = 0u; i < free_top; i++) {
     if (free_stack[i] == page) {
       spin_unlock_irqrestore(SPIN_PAGE, saved);
-      klogf("MM: double-free @ %lx (ra=%lx)\n", addr,
-            (uintptr_t)__builtin_return_address(0));
+      klogf("MM: double-free @ %lx (ra=%lx)\n", (unsigned long)addr,
+            (unsigned long)(uintptr_t)__builtin_return_address(0));
       stack_backtrace();
       return;
     }
